Add files via upload
This project proposes a design variant of obstacle avoidance , bluetooth control 
and voice control for the robot unit. Further advancement of this technology will 
help differently-abled people who cannot operate a vehicle by vice of ailments, 
transportation purposes, hazardous environment places where human interaction 
might be impossible and so on. Also, the use of sensors would prevent collisions. A 
Bluetooth module is used to create a communication link between the car and 
human voice commands via Android Application. The RF transmitter of the 
module can take human voice commands through the application which will then 
be converted into encoded digital data up to an adequate range of 12 meters from 
the robot. The receiver of the module decodes the input data before feeding it 
to the microcontroller to drive DC motors via motor driver L293D for necessary 
movements. An Arduino UNO which is the brain of our system is programmed 
to read voice commands and respond accordingly. Obstacle detection is done by 
ultrasonic sensor interfaced with the microcontroller. Considering this feature, in the 
future it might prove a milestone in vehicle automation. Further the project can be 
developed using Internet of Things, Artificial Intelligence technology where a user 
can control the car from any corner of the world.